Add size, checksum, and format to digital objects
Closes #37

Added in utility method getXML() to get XML at a URL

The format part of this commit is hackish at the moment and this
relates to #38. At this point, the D1 and GeoLink formats are retrieved
at startup and used later on in the script to map between the two
ontologies.