Allow hasGeometryAsWktLiteral to support POINTS
Per the EML standard, if the spatial extent of the data is entered as a
point, the north bound and south bound will be equal and the east bound
and west bound will also be equal. All four bounds are required though.

What this means it that the Solr index should contain either zero
spatial fields or all four, but they might be describing a polygon or a
point. To determine which case you have can be done by comparing the
equality expressed about (N=S, E=W).

This change catches the POINT case and inserts a wkt POINT instead of a
POLYGON.